requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.8
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
defusedxml==0.7.1
    # via djangorestframework-xml
django==3.2.9
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
django-environ==0.8.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-yasg
djangorestframework-xml==2.0.0
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
packaging==21.3
    # via drf-yasg
pyparsing==3.0.6
    # via packaging
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
ruamel.yaml==0.17.17
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via requests
